+++ src/config.rs
@@ -6,15 +6,15 @@ use serde::Deserialize;
 #[serde(deny_unknown_fields)]
 pub struct CylinderMeshConfig {
     #[serde(default = "CylinderMeshConfig::default_half_length")]
-    pub half_length: f32,
+    pub half_length: f64,
 
     #[serde(default = "CylinderMeshConfig::default_resolution")]
-    pub resolution: f32,
+    pub resolution: f64,
 }
 
 impl CylinderMeshConfig {
-    fn default_half_length() -> f32 { 2.0 }
-    fn default_resolution() -> f32 { 0.05 }
+    fn default_half_length() -> f64 { 2.0 }
+    fn default_resolution() -> f64 { 0.05 }
 }
 
 impl Default for CylinderMeshConfig {
@@ -30,16 +30,16 @@ impl Default for CylinderMeshConfig {
 #[serde(deny_unknown_fields)]
 pub struct TorusMeshConfig {
     #[serde(default = "TorusMeshConfig::default_radius")]
-    pub radius: f32,
+    pub radius: f64,
     #[serde(default = "TorusMeshConfig::default_resolution")]
-    pub resolution: f32,
+    pub resolution: f64,
     #[serde(default = "TorusMeshConfig::default_radial_segments")]
     pub radial_segments: usize,
 }
 
 impl TorusMeshConfig {
-    fn default_radius() -> f32 { 0.01 }
-    fn default_resolution() -> f32 { 0.05 }
+    fn default_radius() -> f64 { 0.01 }
+    fn default_resolution() -> f64 { 0.05 }
     fn default_radial_segments() -> usize { 16usize }
 }
 
@@ -57,7 +57,7 @@ impl Default for TorusMeshConfig {
 #[serde(deny_unknown_fields)]
 pub struct RingMeshConfig {
     #[serde(default = "RingMeshConfig::default_radius")]
-    pub radius: f32,
+    pub radius: f64,
     #[serde(default = "RingMeshConfig::default_longitudes")]
     pub longitudes: usize,
     #[serde(default = "RingMeshConfig::default_latitudes")]
@@ -65,7 +65,7 @@ pub struct RingMeshConfig {
 }
 
 impl RingMeshConfig {
-    fn default_radius() -> f32 { 0.02 }
+    fn default_radius() -> f64 { 0.02 }
     fn default_longitudes() -> usize { 16 }
     fn default_latitudes() -> usize { 16 }
 }
